Fixed DataVersion type (was truncated to 16 bits)
Also fixed printing of version numbers
This fixes WEAV-2721 (though more cleanup is needed).

+++ src/lib/profiles/data-management/Current/MessageDef.h
@@ -82,7 +82,10 @@ enum
     kStatus_IncompatibleDataSchemaVersion = 0x2B,
 };
 
-typedef uint16_t DataVersion;
+// This type was wrong (16bit), so stuff was working by chance.
+// The type is only used in a few places.
+// A typedef like this shuold come with the relative PRIDataVersion define
+typedef uint64_t DataVersion;
 
 typedef uint16_t SchemaVersion;
